requirements.txt
langchain_community
langchain_huggingface
langchain_mistralai
mistralai
faiss-cpu
streamlit
requests
pandas
time